fix unit test args type error

+++ pkg/common/util/v1beta2/testutil/tfjob.go
@@ -50,7 +50,7 @@ func NewTFJobWithCleanupJobDelay(chief, worker, ps int, ttl *int32) *tfv1beta2.T
 	return tfJob
 }
 
-func NewTFJobWithActiveDeadlineSeconds(chief, worker, ps int, ads *int32) *tfv1beta2.TFJob {
+func NewTFJobWithActiveDeadlineSeconds(chief, worker, ps int, ads *int64) *tfv1beta2.TFJob {
 	if chief == 1 {
 		tfJob := NewTFJobWithChief(worker, ps)
 		tfJob.Spec.ActiveDeadlineSeconds = ads
